deb: copy current keys to debian package
We're making this copy so that the debian package directory
(deb/tarsnap-archive-keyring/) can be as small as possible, with
no unnecessary files or directories.  That makes reprodicibility
easier.

When we add packages for other distributions or operating systems,
the source for those packages will similarly be placed in their
own directories.