deb: use gpg --export format
This fixes a problem with Debian 9 Stretch.  In particular,
- gpg2.1 added a "keybox" format (which avoids a keyring being slow if you
  have "more than a few thousand keys" [1].  This is used by default.
- apt-get does not support this new format [2].  However, apt-key *does*
  support it.  At least, `apt-key list` will happily parse the files and tell
  you that it knows about public keys in those files [3].

This made testing difficult: a visual inspection of `apt-key list` told me
that the tarsnap-public-keyring package was working on Stretch with the 2017
keys, but when I tried to use actually use those keys, it claimed it didn't
recognize the key.  Also, this problem only arises if the .deb package was
created on Stretch; a binary package created on an older Debian would not use
the keybox format, and would thus be fine.

[1] https://www.gnupg.org/faq/whats-new-in-2.1.html#keybox

[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=844724 , which led to
adding this to the man page for apt-key:

SUPPORTED KEYRING FILES
       apt-key supports only the binary OpenPGP format (also known as "GPG key
       public ring") in files with the "gpg" extension, not the keybox
       database format introduced in newer gpg(1) versions as default for
       keyring files. Binary keyring files intended to be used with any apt
       version should therefore always be created with gpg --export.

[3] reported: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=852473